+++ Editor/I18n/I18nBase.cs
@@ -114,8 +114,6 @@ namespace KRT.VRCQuestTools.I18n
         internal abstract string DeactivateAvatar { get; }
         internal abstract string IncompatibleForQuest { get; }
         internal abstract string MissingScripts { get; }
-        internal abstract string MissingDynamicBone { get; }
-        internal abstract string RemoveMissing { get; }
         internal abstract string ValidatorAlertsProhibitedShaders(string shaderName, string[] materialNames);
         internal abstract string ValidatorAlertsUnsupportedComponents(string componentName, string objectName);
         internal abstract string ValidatorAlertsVeryPoorPhysBones(int count);