+++ Editor/I18n/I18nEnglish.cs
@@ -114,9 +114,7 @@ namespace KRT.VRCQuestTools.I18n
         // Validations
         internal override string DeactivateAvatar => "Deactivate avatar";
         internal override string IncompatibleForQuest => "On Android build target, you can't upload Quest avatars because this avatar can't be uploaded for Quest. Please deactivate such avatars or switch platfrom back to PC.";
-        internal override string MissingScripts => "You can't upload an avatar which has \"missing\" scripts. Please remove these components to build.";
-        internal override string MissingDynamicBone => "Dynamic Bone is missing in the project. Please import Dynamic Bone or remove \"missing\" components to build.";
-        internal override string RemoveMissing => "Remove missing components";
+        internal override string MissingScripts => "There are \"missing\" scripts. Please check for assets or packages you forgot to import.";
         internal override string ValidatorAlertsProhibitedShaders(string shaderName, string[] materialNames) =>
             $"Shader \"{shaderName}\" (Materials: {string.Join(", ", materialNames)}) is not allowed for Quest.";
         internal override string ValidatorAlertsUnsupportedComponents(string componentName, string objectName) =>